Create skeleton for page styling
