+++ src/pages/AboutUs/OurTeam/Cell/Person/personCard.css
@@ -1,0 +1,31 @@
+.person-card {
+  width: 80%;
+  max-width: 700px;
+  display: flex;
+  flex-direction: column;
+  align-items: left;
+}
+
+.person-position {
+  width: 100%;
+  max-width: 700px;
+  margin-top: 30px;
+  font-size: 26px;
+  font-weight: 700;
+}
+
+.person-name {
+  width: 100%;
+  max-width: 700px;
+  margin-top: 20px;
+  font-size: 22px;
+  font-weight: 700;
+}
+
+.person-email {
+  width: 100%;
+  max-width: 700px;
+  margin-top: 10px;
+  font-size: 18px;
+  font-weight: 400;
+}
