+++ src/pages/AboutUs/OurTeam/Cell/cellCard.css
@@ -1,0 +1,24 @@
+.cell-card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  padding: 2px 0px;
+  margin-top: 10px;
+  margin-bottom: 20px;
+}
+
+.cell-name {
+  width: 80%;
+  max-width: 700px;
+  font-size: 40px;
+  font-weight: 400;
+
+  padding: 2px 0px;
+  margin-bottom: 8px;
+}
+
+.cell-image {
+  width: 80%;
+  max-width: 700px;
+}
