+++ src/pages/AboutUs/OurTeam/index.jsx
@@ -1,7 +1,7 @@
 import { useFetchData } from "../../../common/useFetchData";
 import { Routes } from "../../../constants/routes";
 import "./ourteam.css";
-import PersonCard from "./components/PersonCard";
+import CellCard from "./Cell/CellCard";
 import Navbar from "../../Navbar";
 import Footer from "../../Footer";
 import { extractPersonsData } from "./helper";
@@ -10,14 +10,13 @@ const OurTeam = () => {
   const url = Routes.backendRoot + Routes.ourTeam;
   const [isLoading, content] = useFetchData(url);
   const personsData = extractPersonsData(content);
-  console.log(personsData)
   return (
     <div>
       <Navbar />
     <div className="wrapper">
     {isLoading 
       ? <div></div>
-    : personsData.map((item) => <PersonCard name={item.name} position={item.position} email={item.email} emailHyperlink={item.emailHyperlink}/>)}
+      : personsData.map((item) => <CellCard cellName={item.cellName} image={item.image} members={[...item.members]}/>)}
     </div>
     <Footer />
     </div>
