+++ src/pages/AboutUs/OurTeam/ourteam.css
@@ -1,5 +1,5 @@
 .wrapper {
   display: flex;
-  flex-flow: row wrap;
+  flex-direction: column;
   justify-content: center;
 }
